Player YAY!

+++ src/com/onemorepoint/novacraft/game/Game.java
@@ -1,11 +1,12 @@
 package com.onemorepoint.novacraft.game;
 
+import com.onemorepoint.novacraft.*;
 import javax.microedition.khronos.opengles.GL10;
 import java.lang.System;
 
 public class Game
 {
-	protected double elapsedTime;
+	protected float elapsedTime;
 	
 	private long lastTime;
 	private long thisTime;	
@@ -23,7 +24,7 @@ public class Game
 	{
 		lastTime = thisTime;
 		thisTime = System.nanoTime();
-		elapsedTime = (double)(thisTime - lastTime) * 0.000000001;
+		elapsedTime = (float)((thisTime - lastTime) * 0.000000001);
 	}
 		
 	public void Initialize()
